+++ packages/common/src/types/savedCharts.ts
@@ -51,11 +51,6 @@ export type BigNumber = {
     comparisonLabel?: string;
 };
 
-export type BigNumberConfig = {
-    type: ChartType.BIG_NUMBER;
-    config?: BigNumber;
-};
-
 export const PieChartValueLabels = {
     hidden: 'Hidden',
     inside: 'Inside',
@@ -95,11 +90,6 @@ export type PieChart = {
     legendPosition?: PieChartLegendPosition;
 };
 
-export type PieChartConfig = {
-    type: ChartType.PIE;
-    config?: PieChart;
-};
-
 export type ColumnProperties = {
     visible?: boolean;
     name?: string;
@@ -115,11 +105,6 @@ export type TableChart = {
     columns?: Record<string, ColumnProperties>;
     conditionalFormattings?: ConditionalFormattingConfig[];
     metricsAsRows?: boolean;
-};
-
-type TableChartConfig = {
-    type: ChartType.TABLE;
-    config?: TableChart;
 };
 
 export enum CartesianSeriesType {
@@ -253,24 +238,47 @@ export type CompleteCartesianChartLayout = {
 
 export type CartesianChartLayout = Partial<CompleteCartesianChartLayout>;
 
+// TODO: temporary type, wanted to avoid {}
+export type CustomVis = {
+    code?: string;
+};
+
 export type CartesianChart = {
     layout: CartesianChartLayout;
     eChartsConfig: EChartsConfig;
 };
 
-export type CartesianChartConfig = {
-    type: ChartType.CARTESIAN;
-    config: CartesianChart;
+export type BigNumberConfig = {
+    type: ChartType.BIG_NUMBER;
+    config?: BigNumber;
 };
 
-export type CustomVisConfig = { config: {}; type: ChartType.CUSTOM };
+export type CartesianChartConfig = {
+    type: ChartType.CARTESIAN;
+    config?: CartesianChart;
+};
+
+export type CustomVisConfig = {
+    type: ChartType.CUSTOM;
+    config?: CustomVis;
+};
+
+export type PieChartConfig = {
+    type: ChartType.PIE;
+    config?: PieChart;
+};
+
+export type TableChartConfig = {
+    type: ChartType.TABLE;
+    config?: TableChart;
+};
 
 export type ChartConfig =
-    | PieChartConfig
     | BigNumberConfig
-    | TableChartConfig
     | CartesianChartConfig
-    | CustomVisConfig;
+    | CustomVisConfig
+    | PieChartConfig
+    | TableChartConfig;
 
 export type SavedChartType = ChartType;
 
@@ -408,9 +416,9 @@ export const getCustomLabelsFromColumnProperties = (
         : undefined;
 
 export const getCustomLabelsFromTableConfig = (
-    config: ChartConfig['config'],
+    config: ChartConfig['config'] | undefined,
 ): Record<string, string> | undefined =>
-    isTableChartConfig(config)
+    config && isTableChartConfig(config)
         ? getCustomLabelsFromColumnProperties(config.columns)
         : undefined;