+++ packages/frontend/src/components/DashboardTiles/DashboardChartTile.tsx
@@ -40,7 +40,7 @@ import { ExportToGoogleSheet } from '../../features/export';
 import useDashboardChart from '../../hooks/dashboard/useDashboardChart';
 import useDashboardFiltersForTile from '../../hooks/dashboard/useDashboardFiltersForTile';
 import useDashboardStorage from '../../hooks/dashboard/useDashboardStorage';
-import { EChartSeries } from '../../hooks/echarts/useEcharts';
+import { EChartSeries } from '../../hooks/echarts/useEchartsCartesianConfig';
 import { uploadGsheet } from '../../hooks/gdrive/useGdrive';
 import useToaster from '../../hooks/toaster/useToaster';
 import { getExplorerUrlFromCreateSavedChartVersion } from '../../hooks/useExplorerRoute';
@@ -186,8 +186,7 @@ const ValidDashboardChartTile: FC<{
 
     return (
         <VisualizationProvider
-            chartType={chart.chartConfig.type}
-            initialChartConfig={chart.chartConfig}
+            chartConfig={chart.chartConfig}
             initialPivotDimensions={chart.pivotConfig?.columns}
             resultsData={resultData}
             explore={explore}
@@ -232,8 +231,7 @@ const ValidDashboardChartTileMinimal: FC<{
     return (
         <VisualizationProvider
             minimal
-            chartType={chart.chartConfig.type}
-            initialChartConfig={chart.chartConfig}
+            chartConfig={chart.chartConfig}
             initialPivotDimensions={chart.pivotConfig?.columns}
             resultsData={resultData}
             isLoading={false}