+++ packages/frontend/src/components/LightdashVisualization/VisualizationConfigCartesian.tsx
@@ -1,0 +1,73 @@
+import { ChartType } from '@lightdash/common';
+import { FC, useEffect } from 'react';
+import useCartesianChartConfig, {
+    CartesianTypeOptions,
+} from '../../hooks/cartesianChartConfig/useCartesianChartConfig';
+import {
+    VisualizationConfig,
+    VisualizationConfigCommon,
+} from './VisualizationProvider';
+
+export type VisualizationConfigCartesian = {
+    chartType: ChartType.CARTESIAN;
+    chartConfig: ReturnType<typeof useCartesianChartConfig>;
+};
+
+export const isCartesianVisualizationConfig = (
+    visualizationConfig: VisualizationConfig | undefined,
+): visualizationConfig is VisualizationConfigCartesian => {
+    return visualizationConfig?.chartType === ChartType.CARTESIAN;
+};
+
+type VisualizationCartesianConfigProps =
+    VisualizationConfigCommon<VisualizationConfigCartesian> & {
+        stacking: boolean | undefined;
+        cartesianType: CartesianTypeOptions | undefined;
+        columnOrder: string[];
+        validPivotDimensions: string[] | undefined;
+        setPivotDimensions: React.Dispatch<
+            React.SetStateAction<string[] | undefined>
+        >;
+    };
+
+const VisualizationCartesianConfig: FC<VisualizationCartesianConfigProps> = ({
+    explore,
+    resultsData,
+    validPivotDimensions,
+    columnOrder,
+    setPivotDimensions,
+    initialChartConfig,
+    onChartConfigChange,
+    stacking,
+    cartesianType,
+    children,
+}) => {
+    const cartesianConfig = useCartesianChartConfig({
+        initialChartConfig,
+        pivotKeys: validPivotDimensions,
+        resultsData,
+        setPivotDimensions,
+        columnOrder,
+        explore,
+        stacking,
+        cartesianType,
+    });
+
+    useEffect(() => {
+        if (!onChartConfigChange || !cartesianConfig.validConfig) return;
+
+        onChartConfigChange({
+            type: ChartType.CARTESIAN,
+            config: cartesianConfig.validConfig,
+        });
+    }, [cartesianConfig, onChartConfigChange]);
+
+    return children({
+        visualizationConfig: {
+            chartType: ChartType.CARTESIAN,
+            chartConfig: cartesianConfig,
+        },
+    });
+};
+
+export default VisualizationCartesianConfig;