+++ packages/frontend/src/components/VisualizationConfigs/ChartConfigPanel/AxesOptions.tsx
@@ -20,6 +20,7 @@ import { FC, useCallback } from 'react';
 import { useToggle } from 'react-use';
 import { useTracking } from '../../../providers/TrackingProvider';
 import { EventName } from '../../../types/Events';
+import { isCartesianVisualizationConfig } from '../../LightdashVisualization/VisualizationConfigCartesian';
 import { useVisualizationContext } from '../../LightdashVisualization/VisualizationProvider';
 
 interface MinMaxProps {
@@ -85,19 +86,21 @@ type Props = {
 };
 
 const AxesOptions: FC<Props> = ({ items }) => {
+    const { visualizationConfig } = useVisualizationContext();
+
+    if (!isCartesianVisualizationConfig(visualizationConfig)) return null;
+
     const {
-        cartesianConfig: {
-            dirtyLayout,
-            dirtyEchartsConfig,
-            setXAxisName,
-            setYAxisName,
-            setYMinValue,
-            setYMaxValue,
-            setShowGridX,
-            setShowGridY,
-            setInverseX,
-        },
-    } = useVisualizationContext();
+        dirtyLayout,
+        dirtyEchartsConfig,
+        setXAxisName,
+        setYAxisName,
+        setYMinValue,
+        setYMaxValue,
+        setShowGridX,
+        setShowGridY,
+        setInverseX,
+    } = visualizationConfig.chartConfig;
 
     const xAxisField = items.find(
         (item) => getItemId(item) === dirtyLayout?.xField,