+++ packages/frontend/src/components/VisualizationConfigs/ChartConfigPanel/index.tsx
@@ -5,17 +5,23 @@ import {
     COLLAPSABLE_CARD_POPOVER_PROPS,
 } from '../../common/CollapsableCard';
 import MantineIcon from '../../common/MantineIcon';
+import { isCartesianVisualizationConfig } from '../../LightdashVisualization/VisualizationConfigCartesian';
 import { useVisualizationContext } from '../../LightdashVisualization/VisualizationProvider';
 import ChartConfigTabs from './ChartConfigTabs';
 
 const ChartConfigPanel: React.FC = () => {
-    const { resultsData, explore, cartesianConfig } = useVisualizationContext();
+    const { resultsData, explore, visualizationConfig } =
+        useVisualizationContext();
+
+    if (!isCartesianVisualizationConfig(visualizationConfig)) return null;
+
+    const { chartConfig } = visualizationConfig;
 
     const disabled =
         !resultsData ||
         resultsData?.rows.length === 0 ||
         !explore ||
-        !cartesianConfig.validCartesianConfig;
+        !chartConfig.validConfig;
 
     return (
         <Popover