+++ packages/frontend/src/hooks/tableVisualization/useTableConfig.ts
@@ -394,7 +394,7 @@ const useTableConfig = (
         [],
     );
 
-    const validTableConfig: TableChart = useMemo(
+    const validConfig: TableChart = useMemo(
         () => ({
             showColumnCalculation,
             showRowCalculation,
@@ -420,7 +420,7 @@ const useTableConfig = (
     return {
         selectedItemIds,
         columnOrder,
-        validTableConfig,
+        validConfig,
         showColumnCalculation,
         setShowColumnCalculation,
         showRowCalculation,