+++ packages/frontend/src/pages/Dashboard.tsx
@@ -6,7 +6,8 @@ import {
     DashboardTileTypes,
     isDashboardChartTileType,
 } from '@lightdash/common';
-import { useProfiler } from '@sentry/react';
+import { captureException, useProfiler } from '@sentry/react';
+
 import React, {
     FC,
     memo,
@@ -41,6 +42,7 @@ import {
 } from '../hooks/dashboard/useDashboard';
 import useDashboardStorage from '../hooks/dashboard/useDashboardStorage';
 import { useOrganization } from '../hooks/organization/useOrganization';
+import useToaster from '../hooks/toaster/useToaster';
 import { deleteSavedQuery } from '../hooks/useSavedQuery';
 import { useSpaceSummaries } from '../hooks/useSpaces';
 import {
@@ -165,6 +167,8 @@ const Dashboard: FC = () => {
     );
     const oldestCacheTime = useDashboardContext((c) => c.oldestCacheTime);
 
+    const { showToastError } = useToaster();
+
     const { data: organization } = useOrganization();
     const hasTemporaryFilters = useMemo(
         () =>
@@ -200,37 +204,50 @@ const Dashboard: FC = () => {
 
     const { tiles: savedTiles } = dashboard || {};
     useEffect(() => {
-        // TODO: The logic in this useEffect isn't right. It's checking if
-        // there are saved tiles, then checking for unsaved tiles,
-        // then replacing the saved tiles with the unsaved ones if they exist.
         if (savedTiles) {
-            // TODO: maybe this should move in the future, but it makes
-            // some sense here since this useEffect is essentially handling
-            // sessions storage
             clearIsEditingDashboardChart();
             const unsavedDashboardTilesRaw = sessionStorage.getItem(
                 'unsavedDashboardTiles',
             );
             sessionStorage.removeItem('unsavedDashboardTiles');
-            let unsavedDashboardTiles = undefined;
             if (unsavedDashboardTilesRaw) {
                 try {
-                    unsavedDashboardTiles = JSON.parse(
+                    const unsavedDashboardTiles = JSON.parse(
                         unsavedDashboardTilesRaw,
                     );
+                    // If there are unsaved tiles, add them to the dashboard
+                    setDashboardTiles((old = []) => {
+                        return [...old, ...unsavedDashboardTiles];
+                    });
+                    setHaveTilesChanged(!!unsavedDashboardTiles);
                 } catch {
-                    // do nothing
+                    showToastError({
+                        title: 'Error parsing chart',
+                        subtitle: 'Unable to save chart in dashboard',
+                    });
+                    console.error(
+                        'Error parsing chart in dashboard. Attempted to parse: ',
+                        unsavedDashboardTilesRaw,
+                    );
+                    captureException(
+                        `Error parsing chart in dashboard. Attempted to parse: ${unsavedDashboardTilesRaw} `,
+                    );
+                }
+            } else {
+                // If there are no dashboard tiles, set them to the saved ones
+                // This is the first time the dashboard is being loaded.
+                if (!dashboardTiles) {
+                    setDashboardTiles(savedTiles);
                 }
             }
-
-            setDashboardTiles(unsavedDashboardTiles || savedTiles);
-            setHaveTilesChanged(!!unsavedDashboardTiles);
         }
     }, [
         setHaveTilesChanged,
         setDashboardTiles,
+        dashboardTiles,
         savedTiles,
         clearIsEditingDashboardChart,
+        showToastError,
     ]);
 
     useEffect(() => {